fix: BSS-safe CLI command split (#333, #301)
Splits cli-commands.hpp into modular command handlers without BSS regression.

## Changes
- Split 2,043-line cli-commands.hpp into 12 command modules
- Reduced main file to 162 lines (92% reduction)
- Created header/implementation pairs for all command categories:
  - help-commands: help, quit
  - button-commands: button press simulation
  - display-commands: mirror, captions, display
  - device-commands: list, select, add, reboot, role, state
  - cable-commands: serial cable connections
  - network-commands: ESP-NOW peer communication
  - game-commands: Konami, games, progress, colors, difficulty, demo
  - stats-commands: FDN game statistics
  - duel-commands: duel history and rematch
  - debug-commands: debugging utilities
  - command-utils: shared utility functions

## BSS-Safe Approach
- Headers contain ONLY declarations (no inline implementations)
- All implementations in separate .cpp files
- Utility functions in dedicated command-utils module
- Prevents static variable duplication across translation units

## Binary Size Impact
- BSS: 4,155,782 bytes (0% change - no regression!)
- text: 1,185,184 bytes (unchanged)
- data: 1,773,667 bytes (unchanged)

## Testing
- ✅ 392/392 native core tests pass
- ✅ 400/400 CLI tests pass
- ✅ ESP32-S3 release build successful
- ✅ CLI simulator builds and runs

Fixes #333
Closes #301
