+++ include/cli/commands/stats-commands.hpp
@@ -1,0 +1,35 @@
+#pragma once
+
+#ifdef NATIVE_BUILD
+
+#include <vector>
+#include <string>
+#include "cli/commands/command-result.hpp"
+#include "cli/cli-device.hpp"
+
+namespace cli {
+
+/**
+ * Game statistics command handlers.
+ */
+class StatsCommands {
+public:
+    /**
+     * Main stats command dispatcher.
+     */
+    static CommandResult cmdStats(const std::vector<std::string>& tokens,
+                                  std::vector<DeviceInstance>& devices,
+                                  int selectedDevice);
+
+private:
+    static CommandResult cmdStatsSummary(DeviceInstance& dev, int deviceIndex);
+    static CommandResult cmdStatsDetailed(DeviceInstance& dev, int deviceIndex);
+    static CommandResult cmdStatsStreaks(DeviceInstance& dev, int deviceIndex);
+    static CommandResult cmdStatsReset(const std::vector<std::string>& tokens,
+                                       std::vector<DeviceInstance>& devices,
+                                       int selectedDevice);
+};
+
+} // namespace cli
+
+#endif // NATIVE_BUILD
